third_party/WebKit/LayoutTests/inspector/sources/debugger/frameworks-dom-xhr-event-breakpoints-expected.txt

Tests framework black-boxing on DOM, XHR and Event breakpoints.



Running: testDOMBreakpoint
Call stack:
  * 0) Framework.appendChild (framework.js:72)
    1) appendElement (frameworks-dom-xhr-event-breakpoints.html:14)
    2)  (:1)
Paused on a "Subtree Modified" breakpoint set on div#rootElement, because a new child was added to that node.

Running: testXHRBreakpoint
Call stack:
  * 0) Framework.sendXHR (framework.js:79)
    1) sendXHR (frameworks-dom-xhr-event-breakpoints.html:19)
    2)  (:1)
Paused on a XMLHttpRequest.

Running: testEventListenerBreakpoint
Call stack:
    0) testElementClicked (frameworks-dom-xhr-event-breakpoints.html:26)
  * 1) Framework_bound (framework.js:105)
  * 2) Framework_eventListener (framework.js:86)
    3) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.html:37)
    4)  (:1)

Executing Resume...
Call stack:
    0) testElementClicked (frameworks-dom-xhr-event-breakpoints.html:26)
  * 1) Framework_bound (framework.js:105)
  * 2) Framework_eventListener (framework.js:86)
    3) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.html:43)
    4)  (:1)

Executing Resume...

Running: testSteppingThroughEventListenerBreakpoint
Executing StepOver...
Call stack:
    0) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.html:33)
    1)  (:1)

Executing StepOver...
Call stack:
    0) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.html:34)
    1)  (:1)

Executing StepOver...
Executing StepOver...
Executing StepOver...
Call stack:
    0) testElementClicked (frameworks-dom-xhr-event-breakpoints.html:26)
  * 1) Framework_bound (framework.js:105)
  * 2) Framework_eventListener (framework.js:86)
    3) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.html:37)
    4)  (:1)

Executing StepOut...
Executing StepOver...
Executing StepOver...
Executing StepOver...
Executing StepOver...
Call stack:
    0) testElementClicked (frameworks-dom-xhr-event-breakpoints.html:26)
  * 1) Framework_bound (framework.js:105)
  * 2) Framework_eventListener (framework.js:86)
    3) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.html:43)
    4)  (:1)

Executing StepOver...
Executing StepOver...
Call stack:
    0) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.html:44)
    1)  (:1)

Executing Resume...

Running: tearDown
